fix: ANT-2634 - Table cell: empty tag should not appears during cell re-rendering
