+++ src/shared/const/classesConst.ts
@@ -1,0 +1,6 @@
+export const COMMON_HEADER_CLASSES = 'px-1 py-0.5 text-left font-semibold';
+
+export const RESIZER_CLASSES =
+  'absolute top-0 h-full w-0.5 cursor-col-resize touch-none select-none bg-gray-900 bg-opacity-50 opacity-0 group-hover:opacity-100';
+
+export const ROW_CLASSES = '[&_tr]:border-b [&_tr]:border-gray-400 [&_tr]:text-body-s';
