+++ src/shared/types/Table.type.ts
@@ -1,0 +1,2 @@
+export type ColumnSizeType = 'pixels' | 'meta';
+export type ColumnResizeMode = 'onChange' | 'onEnd';
